+++ xbmc/games/controllers/windows/ControllerInstaller.cpp
@@ -100,7 +100,7 @@ void CControllerInstaller::Process()
     pProgressDialog->SetLine(0, CVariant{ progressText });
 
     // Set dialog percentage
-    const unsigned int percentage = 100 * (installedCount + 1) / installableAddons.size();
+    const unsigned int percentage = 100 * (installedCount + 1) / static_cast<unsigned int>(installableAddons.size());
     pProgressDialog->SetPercentage(percentage);
 
     if (!CAddonInstaller::GetInstance().InstallOrUpdate(addon->ID(), false, false))
